requirements.txt
akinator
apscheduler
aiofiles
openai
speechrecognition
speedtest-cli
gtts
wget
aiohttp
bs4
enhancer
heroku3
gitpython
google-api-python-client
htmlwebshot
lottie
lxml
numpy>=1.21.2
oauth2client
opencv-python-headless
pillow==8.3.2
profanitydetector
psutil
pypdf2>=1.26.0
pytz
qrcode
requests
tabulate
git+https://github.com/xditya/telegraph.git@41fd800#egg=telegraph
tgcrypto
youtube-search-python
yt-dlp
https://github.com/New-dev0/Telethon/archive/platy.zip
python-decouple
python-dotenv
py-Ayra==2.1.6
pytgcalls==3.0.0.dev24
emoji
gpytranslate
Markdown
ffmpeg-python
mediainfo
pymediainfo